added edit profile page with avatar, name, email,. phoneNumber, birthDate

+++ EmergencyAppSignUp/app/src/main/java/com/lifeSavers/emergencyappsignup/LogInActivity.kt
@@ -136,7 +136,7 @@ class LogInActivity : AppCompatActivity() {
 
                     val postListener = object : ValueEventListener {
                         override fun onDataChange(dataSnapshot: DataSnapshot) {
-                            val userType = dataSnapshot.getValue(Integer::class.java)
+                            val userType = dataSnapshot.getValue(Long::class.java)
 
                             if (userType.toString() == "0") {
                                 startActivity(Intent(this@LogInActivity, MainActivity::class.java))
